+++ AssemblyLine/ComfyLib/Extensions/UnityExtensions.cs
@@ -1,0 +1,29 @@
+﻿namespace ComfyLib;
+
+using System;
+using System.Collections.Generic;
+
+public static class ChatExtensions {
+  public static void AddMessage(this Chat chat, object obj) {
+    if (chat) {
+      chat.AddString($"{obj}");
+      chat.m_hideTimer = 0f;
+    }
+  }
+}
+
+public static class ObjectExtensions {
+  public static T FirstByNameOrThrow<T>(this IEnumerable<T> unityObjects, string name) where T : UnityEngine.Object {
+    foreach (T unityObject in unityObjects) {
+      if (unityObject.name == name) {
+        return unityObject;
+      }
+    }
+
+    throw new InvalidOperationException($"Could not find Unity object of type {typeof(T)} with name: {name}");
+  }
+
+  public static T Ref<T>(this T unityObject) where T : UnityEngine.Object {
+    return unityObject ? unityObject : default;
+  }
+}