+++ AssemblyLine/Patches/InventoryPatch.cs
@@ -1,21 +1,19 @@
-﻿using HarmonyLib;
+﻿namespace AssemblyLine;
 
-using static AssemblyLine.Patches.InventoryGuiPatch;
+using HarmonyLib;
 
-namespace AssemblyLine.Patches {
-  [HarmonyPatch(typeof(Inventory))]
-  internal class InventoryPatch {
-    [HarmonyPostfix]
-    [HarmonyPatch(nameof(Inventory.Changed))]
-    public static void OnInventoryChangedPostfix(Inventory __instance) {
-      if (Player.m_localPlayer == null 
-          || Player.m_localPlayer.GetInventory() != __instance 
-          || !InventoryGui.IsVisible()) {
+[HarmonyPatch(typeof(Inventory))]
+static class InventoryPatch {
+  [HarmonyPostfix]
+  [HarmonyPatch(nameof(Inventory.Changed))]
+  static void OnInventoryChangedPostfix(Inventory __instance) {
+    if (Player.m_localPlayer == null
+        || Player.m_localPlayer.GetInventory() != __instance
+        || !InventoryGui.IsVisible()) {
 
-        return;
-      }
-
-      SetRequirementText();
+      return;
     }
+
+    CraftingManager.SetRequirementText();
   }
 }