+++ AssemblyLine/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 1.2.0
+
+  * Fixed for the `v0.219.14` patch.
+  * Bumped up `<LangVersion>` to C# 12.
+  * Code clean-up and refactoring.
+
 ### 1.1.1
 
   * Fixed bug where crafting not stopped when inventory closed, sometimes continuing to craft items when next opened.
@@ -12,7 +18,8 @@
 
   * Updated for patch 0.217.22. Updated to TMP Text fields
   * Fixed bug where setting max craft amount set count to 0. Count will always be a positive integer.
-  * Fixed bug where increment and decrement buttons non-interactable after closing InventoryGUI on upgrade tab and reopening.
+  * Fixed bug where increment and decrement buttons non-interactable after closing InventoryGUI on upgrade tab and
+    reopening.
   * Fixed bug where crafting continued when InventoryGUI opened after closing without cancelling crafting.
 
 ### 1.0.0